refactor: StaticResourceUtil 내 메서드 별도 객체로 이동 및 기타 리팩토링
