+++ tomcat/src/main/java/org/apache/coyote/http11/controller/StaticResourceHandler.java
@@ -1,14 +1,19 @@
 package org.apache.coyote.http11.controller;
 
+import ch.qos.logback.core.util.FileUtil;
 import org.apache.coyote.http11.model.HttpRequest;
 import org.apache.coyote.http11.model.HttpResponse;
 import org.apache.coyote.http11.model.StatusCode;
 
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.Objects;
-import org.apache.coyote.http11.util.StaticResourceUtil;
 
 public class StaticResourceHandler {
 
+    private static final String STATIC_PREFIX = "static";
+    private static final String DEFAULT_EXTENSION = ".html";
+    private static final String EXTENSION_DELIMITER = ".";
     public static final String DEFAULT_CONTENT_TYPE = "text/html;charset=utf-8";
 
     public void execute(HttpRequest request, HttpResponse response) {
@@ -39,13 +44,33 @@ public class StaticResourceHandler {
             return;
         }
 
-        String body = StaticResourceUtil.getStaticResource(request.getPath());
+        String body = getStaticResource(request.getPath());
         response.setStatusCode(StatusCode.OK);
 
         if (body == null) {
             response.setStatusCode(StatusCode.NOT_FOUND);
-            body = StaticResourceUtil.getStaticResource("/404.html");
+            body = getStaticResource("/404.html");
         }
-        response.setBodyAndContentLength(body);
+        response.setBodyAndContentLength(Objects.requireNonNull(body));
+    }
+
+    public static String getStaticResource(final String resourcePath) {
+        final var wholeResourcePath = getWholeResourcePath(resourcePath);
+
+        try (final var inputStream = FileUtil.class.getClassLoader().getResourceAsStream(wholeResourcePath)) {
+            if (inputStream == null) {
+                return null;
+            }
+            return new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private static String getWholeResourcePath(final String resourcePathPart) {
+        if (resourcePathPart.contains(EXTENSION_DELIMITER)) {
+            return STATIC_PREFIX + resourcePathPart;
+        }
+        return STATIC_PREFIX + resourcePathPart + DEFAULT_EXTENSION;
     }
 }
